refactor(analyse): move filtering functions
- Moved geometric functions from commands/analyse/analyse.py to
commands/analyse/utils/geometry.py.
- Updated analyse import statements.
